Employee data file is now closed

+++ src/pinmanager/tests/test_pinmanager.py
@@ -9,22 +9,9 @@ from pinmanager import PinManager
 
 
 class TestPinManager(unittest.TestCase):
-    def setUp(self):
-        dummydata = {
-            "employees": [
-                {
-                    "PIN_hash": "d40da9e902a33e5fbe7908fc10f9ce144b66b05c994f43ea488779b19a528bfc05f29e6b8f9947ad306fc4a31bf0f0bfd192f6c4d8578c9158142c433a517ebe",
-                    "name": "Default Administrator PIN",
-                    "has_admin": True
-                }
-            ]
-        }
-
-        with open("test_temp_data/dummydata.json", "w") as f:
-            json.dump(dummydata, f)
-
     def tearDown(self):
-        os.remove("test_temp_data/dummydata.json")
+        if os.path.exists("test_temp_data/dummydata.json"):
+            os.remove("test_temp_data/dummydata.json")
 
     def test_0_pinmanager_initialisation(self):
         try:
